extensions func moved to common
